BC-1777 - move logic to uc
